+++ debug_toolbar/panels/history/views.py
@@ -1,12 +1,13 @@
 from django.http import HttpResponseBadRequest, JsonResponse
 from django.template.loader import render_to_string
 
-from debug_toolbar.decorators import require_show_toolbar
+from debug_toolbar.decorators import render_with_toolbar_language, require_show_toolbar
 from debug_toolbar.panels.history.forms import HistoryStoreForm
 from debug_toolbar.toolbar import DebugToolbar
 
 
 @require_show_toolbar
+@render_with_toolbar_language
 def history_sidebar(request):
     """Returns the selected debug toolbar history snapshot."""
     form = HistoryStoreForm(request.GET)
@@ -37,6 +38,7 @@ def history_sidebar(request):
 
 
 @require_show_toolbar
+@render_with_toolbar_language
 def history_refresh(request):
     """Returns the refreshed list of table rows for the History Panel."""
     form = HistoryStoreForm(request.GET)