+++ debug_toolbar/views.py
@@ -1,16 +1,13 @@
 from django.http import JsonResponse
 from django.utils.html import escape
-from django.utils.translation import (
-    get_language,
-    gettext as _,
-    override as lang_override,
-)
+from django.utils.translation import gettext as _
 
-from debug_toolbar.decorators import require_show_toolbar
+from debug_toolbar.decorators import render_with_toolbar_language, require_show_toolbar
 from debug_toolbar.toolbar import DebugToolbar
 
 
 @require_show_toolbar
+@render_with_toolbar_language
 def render_panel(request):
     """Render the contents of a panel"""
     toolbar = DebugToolbar.fetch(request.GET["store_id"])
@@ -22,9 +19,7 @@ def render_panel(request):
         content = "<p>%s</p>" % escape(content)
         scripts = []
     else:
-        lang = toolbar.config["TOOLBAR_LANGUAGE"] or get_language()
-        with lang_override(lang):
-            panel = toolbar.get_panel_by_id(request.GET["panel_id"])
-            content = panel.content
-            scripts = panel.scripts
+        panel = toolbar.get_panel_by_id(request.GET["panel_id"])
+        content = panel.content
+        scripts = panel.scripts
     return JsonResponse({"content": content, "scripts": scripts})